moving javascript files into their own directory
