+++ bsi.html
@@ -1,4 +1,4 @@
-<script src="src/page-loading/web-components-ready.js"></script>
+<script src="js/page-loading/web-components-ready.js"></script>
 
 <link rel="import" href="../polymer/polymer.html">
 <link rel="import" href="../d2l-fastdom-import/fastdom.html">
