+++ package.json
@@ -7,7 +7,7 @@
     "autoprefix": "postcss -c postcss.config.json",
     "prebuild": "rimraf dist && mkdir dist",
     "build": "npm run build:js && npm run build:sass && npm run build:icons && npm run build:wc",
-    "build:js": "browserify -g uglifyify ./src/index.js > ./dist/bsi.js",
+    "build:js": "browserify -g uglifyify ./js/index.js > ./dist/bsi.js",
     "postbuild:js": "npm run uglify:js",
     "build:sass": "node-sass --output-style expanded ./sass/bsi.scss ./dist/bsi.css && node-sass --output-style expanded ./sass/datagrid/datagrid.scss ./dist/datagrid.css",
     "postbuild:sass": "npm run autoprefix && npm run uglify:css",
